Get the text files to write properly
The image file still isn't working properly, seemingly because we are trimming off a bunch of bytes when we check to make sure we don't write empty buffer on packets that don't utilize the entire buffer.
Debugging comments still need to be removed and testing still needs to be implemented.